# docs/04-api/openapi-day22-uc1-replay.fragment.yaml
openapi: 3.1.0
info:
  title: MyoLab-AI Day 22 UC1 Replay API
  version: 0.1.0
  description: >-
    Deterministic engineering replay contract. It is not a validated clinical
    classifier or a production streaming interface.
paths:
  /v1/uc1/sessions/{session_id}/replays:
    post:
      operationId: createUC1Replay
      summary: Create a deterministic UC1 replay
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplayCreateRequest'
      responses:
        '201':
          description: Replay created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplaySession'
              examples:
                replay:
                  $ref: '#/components/examples/ReplaySession'
        '409':
          description: Analysis state, ownership, or idempotency conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: Analysis or upstream replay context not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '422':
          description: Invalid request or missing idempotency key
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/uc1/replays/{replay_id}:
    get:
      operationId: getUC1Replay
      summary: Get only the revealed replay state
      parameters:
        - $ref: '#/components/parameters/ReplayId'
      responses:
        '200':
          description: Current replay state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplaySession'
        '404':
          description: Replay not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/uc1/replays/{replay_id}/advance:
    post:
      operationId: advanceUC1Replay
      summary: Reveal the next deterministic replay window
      description: >-
        Deterministic dev/test transition endpoint. This is not a production
        streaming API.
      parameters:
        - $ref: '#/components/parameters/ReplayId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplayAdvanceRequest'
      responses:
        '200':
          description: Advanced or already-terminal replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplaySession'
        '404':
          description: Replay not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Stale cursor or revision
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Invalid advance request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/uc1/replays/{replay_id}/feedback:
    post:
      operationId: createUC1ReplayFeedback
      summary: Record human feedback for the exact current window
      parameters:
        - $ref: '#/components/parameters/ReplayId'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
        - name: X-Actor-Role
          in: header
          required: true
          schema:
            type: string
            enum:
              - ktv
              - physician
              - researcher
              - ml_qa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '201':
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
              examples:
                feedback:
                  $ref: '#/components/examples/FeedbackResponse'
        '403':
          description: Actor role is not authorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Replay not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Stale window/revision or idempotency conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Invalid feedback semantics
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    ReplayId:
      name: replay_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
    ReplayCreateRequest:
      type: object
      additionalProperties: false
      required:
        - analysisId
        - scenarioId
      properties:
        analysisId:
          type: string
          minLength: 1
        scenarioId:
          $ref: '#/components/schemas/ScenarioId'
    ReplayAdvanceRequest:
      type: object
      additionalProperties: false
      required:
        - expectedCurrentIndex
        - expectedRevision
      properties:
        expectedCurrentIndex:
          type: integer
          minimum: -1
        expectedRevision:
          type: integer
          minimum: 0
    FeedbackRequest:
      type: object
      additionalProperties: false
      required:
        - action
        - reviewerCertainty
        - expectedWindowId
        - expectedRevision
      properties:
        action:
          enum:
            - accept
            - correct
            - uncertain
            - remeasure
        correctedGesture:
          oneOf:
            - $ref: '#/components/schemas/GestureId'
            - type: 'null'
        reviewerCertainty:
          enum:
            - low
            - moderate
            - high
        expectedWindowId:
          type: string
          minLength: 1
        expectedRevision:
          type: integer
          minimum: 0
    ScenarioId:
      type: string
      enum:
        - uc1_golden_correct
        - uc1_ambiguous_prediction
        - uc1_no_activity
        - uc1_fatigue_confidence_drop
        - uc1_electrode_shift_warning
        - uc1_qc_fail_abstention
        - uc1_device_disconnect
    GestureId:
      type: string
      enum:
        - rest
        - hand_open
        - hand_close
        - wrist_flexion
        - wrist_extension
    ActiveGestureId:
      type: string
      enum:
        - hand_open
        - hand_close
        - wrist_flexion
        - wrist_extension
    EngineeringConfidence:
      type: string
      enum:
        - engineering_high
        - engineering_moderate
        - engineering_low
        - engineering_very_low
        - not_available
    GestureSegmentReference:
      type: object
      additionalProperties: false
      required:
        - rawSignalRef
        - sourceHashSha256
        - startSample
        - endSampleExclusive
        - startTimeS
        - endTimeExclusiveS
        - channelIds
        - repetitionId
        - calibrationId
      properties:
        rawSignalRef:
          type: string
          minLength: 1
        sourceHashSha256:
          $ref: "#/components/schemas/Sha256"
        startSample:
          type: integer
          minimum: 0
        endSampleExclusive:
          type: integer
          minimum: 1
        startTimeS:
          type: number
          minimum: 0
        endTimeExclusiveS:
          type: number
          exclusiveMinimum: 0
        channelIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            minLength: 1
        repetitionId:
          type: string
          minLength: 1
        calibrationId:
          type: string
          minLength: 1
    ActivityGateSummary:
      type: object
      additionalProperties: false
      required:
        - status
        - windowRmsUv
        - activationThresholdUv
        - releaseThresholdUv
        - reasonCode
      properties:
        status:
          type: string
          enum: [active, inactive, uncertain]
        windowRmsUv:
          type: number
          minimum: 0
        activationThresholdUv:
          type: number
          exclusiveMinimum: 0
        releaseThresholdUv:
          type: number
          exclusiveMinimum: 0
        reasonCode:
          type: string
          minLength: 1
    QualityContext:
      type: object
      additionalProperties: false
      required: [source, qualityResultId, status, reasonCodes]
      properties:
        source:
          type: string
          enum:
            - day20_quality_gate
            - day17_signal_quality
            - scenario_fixture
            - not_available
        qualityResultId:
          type: string
          minLength: 1
        status:
          type: string
          enum: [pass, warning, fail]
        reasonCodes:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
    FatigueOverlay:
      type: object
      additionalProperties: false
      required:
        - source
        - status
        - confidenceAdjustmentApplied
        - reasonCodes
        - evidenceSummaryVi
        - counterevidenceVi
        - limitationsVi
      properties:
        source:
          type: string
          enum: [scenario_fixture, analysis_summary, not_available]
        status:
          type: string
          enum: [stable, warning, abstain, not_available]
        confidenceAdjustmentApplied:
          type: boolean
        reasonCodes:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
        evidenceSummaryVi:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
        counterevidenceVi:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
        limitationsVi:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
    LatencyBreakdown:
      type: object
      additionalProperties: false
      required:
        - totalMs
        - acquisitionMs
        - windowMs
        - preprocessMs
        - inferenceMs
        - transportRenderMs
      properties:
        totalMs: {type: number, minimum: 0}
        acquisitionMs: {type: number, minimum: 0}
        windowMs: {type: number, minimum: 0}
        preprocessMs: {type: number, minimum: 0}
        inferenceMs: {type: number, minimum: 0}
        transportRenderMs: {type: number, minimum: 0}
    GestureSafety:
      type: object
      additionalProperties: false
      required:
        - scoreIsProbability
        - clinicalUseAllowed
        - rawSamplesIncluded
        - physicalActuationAllowed
      properties:
        scoreIsProbability: {const: false}
        clinicalUseAllowed: {const: false}
        rawSamplesIncluded: {const: false}
        physicalActuationAllowed: {const: false}
    GestureInferenceWindow:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - windowId
        - sessionId
        - analysisId
        - protocolVersion
        - segmentRef
        - targetGesture
        - activityGate
        - predictedGesture
        - baseEngineeringConfidence
        - engineeringConfidence
        - qualityContext
        - fatigueOverlay
        - deviceState
        - latency
        - requiresHumanReview
        - sourceType
        - modelVersion
        - modelValidationStatus
        - resultHashSha256
        - safety
      properties:
        schemaVersion: {const: gesture-inference.v0.1}
        windowId: {type: string, minLength: 1}
        sessionId: {type: string, minLength: 1}
        analysisId: {type: string, minLength: 1}
        protocolVersion: {type: string, minLength: 1}
        segmentRef:
          $ref: "#/components/schemas/GestureSegmentReference"
        targetGesture:
          $ref: "#/components/schemas/ActiveGestureId"
        activityGate:
          $ref: "#/components/schemas/ActivityGateSummary"
        predictedGesture:
          oneOf:
            - $ref: "#/components/schemas/ActiveGestureId"
            - type: "null"
        baseEngineeringConfidence:
          $ref: "#/components/schemas/EngineeringConfidence"
        engineeringConfidence:
          $ref: "#/components/schemas/EngineeringConfidence"
        qualityContext:
          $ref: "#/components/schemas/QualityContext"
        fatigueOverlay:
          $ref: "#/components/schemas/FatigueOverlay"
        deviceState:
          type: string
          enum: [connected, disconnected, reconnecting]
        latency:
          $ref: "#/components/schemas/LatencyBreakdown"
        requiresHumanReview: {const: true}
        sourceType: {const: synthetic_replay}
        modelVersion: {type: string, minLength: 1}
        modelValidationStatus: {const: not_validated}
        resultHashSha256:
          $ref: "#/components/schemas/Sha256"
        safety:
          $ref: "#/components/schemas/GestureSafety"
    ReplayLatencySummary:
      type: object
      additionalProperties: false
      required:
        - observedWindowCount
        - p50Ms
        - p95Ms
        - droppedWindows
        - disconnectTimeoutMs
      properties:
        observedWindowCount: {type: integer, minimum: 0}
        p50Ms:
          type: [number, "null"]
          minimum: 0
        p95Ms:
          type: [number, "null"]
          minimum: 0
        droppedWindows: {type: integer, minimum: 0}
        disconnectTimeoutMs: {type: integer, exclusiveMinimum: 0}
    ReplaySession:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - replayId
        - sessionId
        - analysisId
        - scenarioId
        - state
        - currentIndex
        - revision
        - totalWindows
        - currentWindow
        - history
        - latencySummary
        - reasonCodes
        - sourceType
        - modelValidationStatus
        - clinicalUseAllowed
        - humanReviewRequired
      properties:
        schemaVersion: {const: uc1-replay-session.v0.1}
        replayId: {type: string, minLength: 1}
        sessionId: {type: string, minLength: 1}
        analysisId: {type: string, minLength: 1}
        scenarioId:
          $ref: "#/components/schemas/ScenarioId"
        state:
          type: string
          enum: [idle, running, completed, abstained, disconnected, failed]
        currentIndex: {type: integer, minimum: -1}
        revision: {type: integer, minimum: 0}
        totalWindows: {type: integer, minimum: 0}
        currentWindow:
          description: Current revealed window; its windowId MUST NOT appear in history.
          oneOf:
            - $ref: "#/components/schemas/GestureInferenceWindow"
            - type: "null"
        history:
          description: Chronological past-only windows; length equals max(0, currentIndex).
          type: array
          items:
            $ref: "#/components/schemas/GestureInferenceWindow"
        latencySummary:
          $ref: "#/components/schemas/ReplayLatencySummary"
        reasonCodes:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
        sourceType: {const: synthetic_replay}
        modelValidationStatus: {const: not_validated}
        clinicalUseAllowed: {const: false}
        humanReviewRequired: {const: true}
    FeedbackResponse:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - feedbackId
        - replayId
        - action
        - reviewerCertainty
        - actorRole
        - context
        - automaticTrainingCandidate
      properties:
        schemaVersion: {const: gesture-feedback.v0.1}
        feedbackId: {type: string, minLength: 1}
        replayId: {type: string, minLength: 1}
        action:
          type: string
          enum: [accept, correct, uncertain, remeasure]
        correctedGesture:
          oneOf:
            - $ref: "#/components/schemas/GestureId"
            - type: "null"
        reviewerCertainty:
          type: string
          enum: [low, moderate, high]
        actorRole:
          type: string
          enum: [ktv, physician, researcher, ml_qa]
        context:
          $ref: "#/components/schemas/FeedbackContext"
        automaticTrainingCandidate: {const: false}
    FeedbackContext:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - analysisId
        - sessionId
        - windowId
        - rawSignalRef
        - sourceHashSha256
        - startSample
        - endSampleExclusive
        - startTimeS
        - endTimeExclusiveS
        - channelIds
        - repetitionId
        - calibrationId
        - modelVersion
        - originalResultHashSha256
      properties:
        schemaVersion:
          const: gesture-feedback-context.v0.1
        analysisId:
          type: string
        sessionId:
          type: string
        windowId:
          type: string
        rawSignalRef:
          type: string
        sourceHashSha256:
          $ref: '#/components/schemas/Sha256'
        startSample:
          type: integer
          minimum: 0
        endSampleExclusive:
          type: integer
          minimum: 1
        startTimeS:
          type: number
          minimum: 0
        endTimeExclusiveS:
          type: number
          exclusiveMinimum: 0
        channelIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
        repetitionId:
          type: string
        calibrationId:
          type: string
        modelVersion:
          type: string
        originalResultHashSha256:
          $ref: '#/components/schemas/Sha256'
    Sha256:
      type: string
      pattern: "^[0-9a-f]{64}$"
    InvalidParameter:
      type: object
      additionalProperties: false
      required: [field, message, code]
      properties:
        field:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
        - error_code
        - trace_id
      properties:
        type: {type: string, minLength: 1}
        title: {type: string, minLength: 1}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string, minLength: 1}
        instance: {type: string, minLength: 1}
        error_code: {type: string, minLength: 1}
        trace_id: {type: string, minLength: 1}
        invalid_params:
          type: array
          items:
            $ref: "#/components/schemas/InvalidParameter"
  examples:
    ReplaySession:
      value:
        schemaVersion: uc1-replay-session.v0.1
        replayId: REPLAY-000000000001
        sessionId: SESSION-D20-001
        analysisId: AN21-000000000001
        scenarioId: uc1_golden_correct
        state: idle
        currentIndex: -1
        revision: 0
        totalWindows: 4
        currentWindow: null
        history: []
        latencySummary:
          observedWindowCount: 0
          p50Ms: null
          p95Ms: null
          droppedWindows: 0
          disconnectTimeoutMs: 2000
        reasonCodes: []
        sourceType: synthetic_replay
        modelValidationStatus: not_validated
        clinicalUseAllowed: false
        humanReviewRequired: true
    FeedbackResponse:
      value:
        schemaVersion: gesture-feedback.v0.1
        feedbackId: FEEDBACK-000000000001
        replayId: REPLAY-000000000001
        action: correct
        correctedGesture: hand_open
        reviewerCertainty: high
        actorRole: ktv
        context:
          schemaVersion: gesture-feedback-context.v0.1
          analysisId: AN21-000000000001
          sessionId: SESSION-D20-001
          windowId: WIN-UC1-0001
          rawSignalRef: RAW-REF-D20-001
          sourceHashSha256: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
          startSample: 5000
          endSampleExclusive: 6000
          startTimeS: 5.0
          endTimeExclusiveS: 6.0
          channelIds:
            - CH01
            - CH02
          repetitionId: REP-D20-001
          calibrationId: CAL-SESSION-D20-001
          modelVersion: gesture-replay-v0.1
          originalResultHashSha256: bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
        automaticTrainingCandidate: false